+++ anzeigen-frontend/src/components/Ad/dialog/AdAgbAccept.vue
