+++ anzeigen-frontend/src/components/Ad/list/AdCategoryChip.vue
