+++ anzeigen-frontend/src/components/Ad/list/AdPrice.vue
