+++ anzeigen-frontend/src/components/filter/FilterAdCategory.vue
